Moved MapOffset.h to libOPHD and Updated includes
