+++ OPHD/Map/MapView.h
@@ -1,7 +1,8 @@
 #pragma once
 
 #include "../Map/MapCoordinate.h"
-#include "../Map/MapOffset.h"
+
+#include <libOPHD/Map/MapOffset.h>
 
 #include <NAS2D/Math/Rectangle.h>
