+++ libOPHD/Map/MapOffset.h
